pyproject.toml
[project]
name = "smartframe"
version = "0.1.0"
description = "AI-powered portrait editing and professional photo formatting tool."
readme = "README.md"
authors = [
    { name = "Kumar", email = "kumar@example.com" }
]
requires-python = ">=3.13"
dependencies = [
    "customtkinter>=5.2.2",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "rembg[cpu]>=2.0.72",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
managed = true
package = false
